refactor the types
