+++ packages/dox-cli/COMMAND_REGISTRY.md
@@ -2,6 +2,12 @@
 
 The Dox CLI has been refactored to use a command registry system that makes it easy to add new commands.
 
+## File Structure
+
+- `lib/src/types.dart` - Contains `CommandCategory` enum and `CommandFunction` typedef
+- `lib/src/command_registry.dart` - Contains `CommandDefinition` class and `CommandRegistry`
+- `lib/src/tools/help.dart` - Help system that uses the command registry
+
 ## How to Add a New Command
 
 ### 1. Define the Command
@@ -12,6 +18,7 @@ CommandRegistry.addCommand(
     command: 'my:command',
     helpInfo: 'my:command <arg1> [arg2]',
     description: 'Description of what this command does',
+    category: CommandCategory.generation, // Use enum from types.dart
     minArgs: 1,        // Minimum number of arguments required
     maxArgs: 3,        // Maximum number of arguments allowed (-1 for unlimited)
     aliases: ['mc', 'my'], // Optional aliases for the command
@@ -32,12 +39,29 @@ CommandRegistry.addCommand(
 - **command**: The main command name (e.g., 'create:controller')
 - **helpInfo**: Usage information shown in help (e.g., 'create:controller <name> [-r]')
 - **description**: Help text for the command
+- **category**: CommandCategory enum value (e.g., CommandCategory.generation)
 - **minArgs**: Minimum number of arguments required
 - **maxArgs**: Maximum number of arguments allowed (-1 for unlimited)
 - **aliases**: Optional list of alternative command names
 - **function**: The actual function that executes the command
 
-### 3. Example Commands
+### 3. Available Categories
+
+The `CommandCategory` enum provides the following categories:
+
+```dart
+enum CommandCategory {
+  project('Project', 0),        // Project creation commands
+  development('Development', 1), // Development server commands
+  build('Build', 2),           // Build commands
+  buildRunner('Build Runner', 3), // Build runner commands
+  generation('Generation', 4),  // Code generation commands
+  database('Database', 5),      // Database/migration commands
+  system('System', 6);          // System utility commands
+}
+```
+
+### 4. Example Commands
 
 Here are some examples of how commands are defined in the registry:
 
@@ -47,6 +71,7 @@ CommandDefinition(
   command: 'help',
   helpInfo: 'help [command]',
   description: 'Show help information',
+  category: CommandCategory.system,
   minArgs: 0,
   maxArgs: 1,
   function: (args) async {
@@ -63,6 +88,7 @@ CommandDefinition(
   command: 'create:model',
   helpInfo: 'create:model <model_name>',
   description: 'Create a new model',
+  category: CommandCategory.generation,
   minArgs: 1,
   maxArgs: 1,
   function: (args) async {
@@ -75,6 +101,7 @@ CommandDefinition(
   command: 'create:controller',
   helpInfo: 'create:controller <controller_name> [-r] [-ws]',
   description: 'Create a new controller',
+  category: CommandCategory.generation,
   minArgs: 1,
   maxArgs: 2,
   function: (args) async {
@@ -89,6 +116,7 @@ CommandDefinition(
   command: 'serve',
   helpInfo: 'serve or s [--ignore-build-runner]',
   description: 'Start the development server',
+  category: CommandCategory.development,
   aliases: ['server', 's'],
   minArgs: 0,
   maxArgs: 1,
@@ -103,7 +131,7 @@ CommandDefinition(
 ),
 ```
 
-### 4. Help System
+### 5. Help System
 
 The new help system automatically generates help information from the CommandDefinition:
 
@@ -112,10 +140,13 @@ The new help system automatically generates help information from the CommandDef
 dox help
 ```
 Shows all commands organized by category:
-- **Development**: serve, build, build_runner commands
-- **Generation**: create:* commands
+- **Project**: create commands
+- **Development**: serve, create:controller, create:middleware, create:request commands
+- **Build**: build commands
+- **Build Runner**: build_runner commands
+- **Generation**: create:model, create:serializer, key:generate commands
 - **Database**: migration commands
-- **System**: version, help, update, key:generate
+- **System**: help, update, version commands
 
 #### Command-Specific Help
 ```bash
@@ -124,6 +155,7 @@ dox help <command>
 Shows detailed information about a specific command including:
 - Usage syntax
 - Description
+- Category
 - Aliases (if any)
 - Argument requirements
 
@@ -137,21 +169,23 @@ Output:
 Command: create:controller
 Usage: dox create:controller <controller_name> [-r] [-ws]
 Description: Create a new controller
+Category: Generation
 Arguments: 1 to 2 arguments
 ```
 
-### 5. Benefits of This System
+### 6. Benefits of This System
 
 1. **Easy to Add**: Just define a CommandDefinition and add it to the registry
-2. **Type Safe**: All commands are strongly typed
+2. **Type Safe**: All commands are strongly typed using enums
 3. **Argument Validation**: Built-in argument count validation
 4. **Aliases Support**: Multiple ways to call the same command
 5. **Centralized**: All commands are defined in one place
 6. **Maintainable**: Clear structure makes it easy to understand and modify
 7. **Auto-Generated Help**: Help information is automatically generated from command definitions
 8. **Categorized**: Commands are organized by category for better UX
+9. **Well Organized**: Types are separated into their own file for better maintainability
 
-### 6. Adding Commands at Runtime
+### 7. Adding Commands at Runtime
 
 You can also add commands dynamically:
 
@@ -162,6 +196,7 @@ CommandRegistry.addCommand(
     command: 'custom:command',
     helpInfo: 'custom:command <arg>',
     description: 'A custom command',
+    category: CommandCategory.system,
     minArgs: 1,
     maxArgs: 1,
     function: (args) async {
