+++ packages/dox-cli/lib/src/version.dart
@@ -1,0 +1,1 @@
+const version = '2.0.2';
